+++ torchbearer/callbacks/tensor_board.py
@@ -114,6 +114,9 @@ class AbstractTensorBoard(Callback):
         comment (str): Descriptive comment to append to path
         visdom (bool): If true, log to visdom instead of tensorboard
         visdom_params (VisdomParams): Visdom parameter settings object, uses default if None
+
+    State Requirements:
+        - :attr:`torchbearer.state.MODEL`: PyTorch model
     """
 
     def __init__(self, log_dir='./logs',
@@ -167,6 +170,22 @@ class AbstractTensorBoard(Callback):
 
     @staticmethod
     def add_metric(add_fn, tag, metric, *args, **kwargs):
+        """ Static method that recurses through `metric` until the `add_fn` can be applied. Useful when metric is an
+        iterable of tensors so that the tensors can  all be passed to an `add_fn` such as writer.add_scalar.
+        For example, if passed `metric` as [[A, B], [C, ], D, {'E': E}] then `add_fn` would be called on A, B, C, D and
+        E and the respective tags (with base tag 'met') would be: met_0_0, met_0_1, met_1_0, met_2, met_E. Throws a
+        warning if `add_fn` fails to parse a metric.
+
+        Args:
+            add_fn: Function to be called to log a metric, e.g. SummaryWriter.add_scalar
+            tag: Tag under which to log the metric
+            metric: Iterable of metrics.
+            *args: Args for `add_fn`
+            **kwargs: Keyword args for `add_fn`
+
+        Returns:
+
+        """
         try:
             add_fn(tag, metric, *args, **kwargs)
         except NotImplementedError:
@@ -188,12 +207,12 @@ class AbstractTensorBoard(Callback):
 class TensorBoard(AbstractTensorBoard):
     """TensorBoard callback which writes metrics to the given log directory. Requires the TensorboardX library for python.
 
-    Example::
+    Example: ::
+
         >>> from torchbearer import Trial
         >>> from torchbearer.callbacks import TensorBoard
         >>> import datetime
         >>> current_time = datetime.now().strftime('%b%d_%H-%M-%S')
-
         # Callback that will log to tensorboard under "(model name)_(current time)"
         >>> tb = TensorBoard(log_dir='./logs', write_graph=False, comment=current_time)
         # Trial that will run the callback and log accuracy and loss metrics
@@ -208,6 +227,14 @@ class TensorBoard(AbstractTensorBoard):
         comment (str): Descriptive comment to append to path
         visdom (bool): If true, log to visdom instead of tensorboard
         visdom_params (VisdomParams): Visdom parameter settings object, uses default if None
+
+    State Requirements:
+        - :attr:`torchbearer.state.MODEL`: PyTorch model
+        - :attr:`torchbearer.state.EPOCH`: State should have the current epoch stored
+        - :attr:`torchbearer.state.X`: State should have the current data stored if a model graph is to be built
+        - :attr:`torchbearer.state.BATCH`: State should have the current batch number stored if logging batch metrics
+        - :attr:`torchbearer.state.TRAIN_STEPS`: State should have the number of training steps stored
+        - :attr:`torchbearer.state.METRICS`: State should have a dictionary of metrics stored
     """
 
     def __init__(self, log_dir='./logs',
@@ -293,12 +320,12 @@ class TensorBoardText(AbstractTensorBoard):
     """TensorBoard callback which writes metrics as text to the given log directory. Requires the TensorboardX library
     for python.
 
-    Example::
+    Example: ::
+
         >>> from torchbearer import Trial
         >>> from torchbearer.callbacks import TensorBoardText
         >>> import datetime
         >>> current_time = datetime.now().strftime('%b%d_%H-%M-%S')
-
         # Callback that will log to tensorboard under "(model name)_(current time)"
         >>> tb = TensorBoardText(comment=current_time)
         # Trial that will run the callback and log accuracy and loss metrics as text to tensorboard
@@ -312,6 +339,12 @@ class TensorBoardText(AbstractTensorBoard):
         comment (str): Descriptive comment to append to path
         visdom (bool): If true, log to visdom instead of tensorboard
         visdom_params (VisdomParams): Visdom parameter settings object, uses default if None
+
+    State Requirements:
+        - :attr:`torchbearer.state.SELF`: The :attr:`torchbearer.Trial` running this callback
+        - :attr:`torchbearer.state.EPOCH`: State should have the current epoch stored
+        - :attr:`torchbearer.state.BATCH`: State should have the current batch number stored if logging batch metrics
+        - :attr:`torchbearer.state.METRICS`: State should have a dictionary of metrics stored
     """
 
     def __init__(self, log_dir='./logs',
@@ -393,16 +426,17 @@ class TensorBoardImages(AbstractTensorBoard):
     TensorboardX library and torchvision.utils.make_grid (requires torchvision). Images are selected from the given key and saved to the given
     path. Full name of image sub directory will be model name + _ + comment.
 
-    Example::
+    Example: ::
+
         >>> from torchbearer import Trial, state_key
         >>> from torchbearer.callbacks import TensorBoardImages
         >>> import datetime
         >>> current_time = datetime.now().strftime('%b%d_%H-%M-%S')
         >>> IMAGE_KEY = state_key('image_key')
 
-        # Callback that will log to tensorboard under "(model name)_(current time)"
+        >>> # Callback that will log to tensorboard under "(model name)_(current time)"
         >>> tb = TensorBoardImages(comment=current_time, name='test_image', key=IMAGE_KEY)
-        # Trial that will run log to tensorboard images stored under IMAGE_KEY
+        >>> # Trial that will run log to tensorboard images stored under IMAGE_KEY
         >>> t = Trial(None, callbacks=[tb], metrics=['acc', 'loss'])
 
     Args:
@@ -420,6 +454,10 @@ class TensorBoardImages(AbstractTensorBoard):
         pad_value: See `torchvision.utils.make_grid <https://pytorch.org/docs/stable/torchvision/utils.html#torchvision.utils.make_grid>`_
         visdom (bool): If true, log to visdom instead of tensorboard
         visdom_params (VisdomParams): Visdom parameter settings object, uses default if None
+
+    State Requirements:
+        - :attr:`torchbearer.state.EPOCH`: State should have the current epoch stored
+        - `key`: State should have images stored under the given state key
     """
 
     def __init__(self, log_dir='./logs',
@@ -509,6 +547,11 @@ class TensorBoardProjector(AbstractTensorBoard):
         write_features (bool): If True, the image features will be written as an embedding
         features_key (StateKey): The key in state to use for the embedding. Typically model output but can be used to show
             features from any layer of the model.
+
+    State Requirements:
+        - :attr:`torchbearer.state.EPOCH`: State should have the current epoch stored
+        - :attr:`torchbearer.state.X`: State should have the current data stored
+        - :attr:`torchbearer.state.Y_TRUE`: State should have the current targets stored
     """
 
     def __init__(self, log_dir='./logs',